// queries/__generated__/update_CreateProfileMutation.graphql.ts
export type PartialProfile = {
  classStanding: string;
  email: string;
  firstName: string;
  graduation: GraduationInputType;
  lastName: string;
  major: string;
  netid: string;
  user: any;
  utdStudent: boolean;
};
export type GraduationInputType = {
  semester: string;
  year: string;
};
export type update_CreateProfileMutation$variables = {
  profile: PartialProfile;
};
export type update_CreateProfileMutation$data = {
  readonly createProfile: {
    readonly email: string;
    readonly firstName: string;
    readonly lastName: string;
    readonly graduation: {
      readonly semester: string;
      readonly year: string;
    };
    readonly classStanding: string;
    readonly _id: any;
    readonly major: string;
    readonly membershipStatus: boolean;
  };
};
export type update_CreateProfileMutation = {
  variables: update_CreateProfileMutation$variables;
  response: update_CreateProfileMutation$data;
};